+++ neoforge21/src/main/java/com/vecoo/extralib/item/UtilItem.java
@@ -36,7 +36,7 @@ public class UtilItem {
             try {
                 itemStack.set(DataComponents.CUSTOM_MODEL_DATA, new CustomModelData(Integer.parseInt(parts[2])));
             } catch (NumberFormatException e) {
-                ExtraLib.getLogger().error("[ExtraLib] Invalid CustomModelData value in item: " + itemId);
+                ExtraLib.getLogger().error("Invalid CustomModelData value in item: " + itemId);
             }
         }