+++ changelog.md
@@ -1,2 +1,5 @@
-* Added annotations.
-* Optimized get UUID.
+* Forge 1.16.5 unsupported.
+* Fixes slf4j warning.
+* Update depends.
+* Improve logger.
+* Optimized gson.